+++ docs/CHANGELOG.md
@@ -6,7 +6,7 @@ A detailed list of all notable changes for every released version.
 
 ## [0.0.89](https://github.com/blooper-gg/toodle/releases/tag/0.0.89)
 
-- Add `JumboQuadNode` to render a jumbo texture, see [jumbo texture example](https://toodle.gg/examples/jumbo-texture.html) for more details.
+- Add `JumboQuadNode` to render a jumbo texture, see [jumbo texture example](https://toodle.gg/examples/jumbo-textures.html) for more details.
 
 ## [0.0.88](https://github.com/blooper-gg/toodle/releases/tag/0.0.88)